Automate the notarization process with bash script
Skpm added support for plugin notarization when publishing a new version but the CLI fails due some string parsing (no update on the CLI since December last year). An alternative to the internal CLI notarization logic is using custom bash script for firing the notarization command and handling the output.